+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Now requiring Ruby 2.4+ [#48], [#51]
 * Removed `coercible` dependency as now all coercion functionality is implemented locally. This is a backwards compatible change. [#49]
+* Calling `ENVied.required?` now returns an explicit `false` or `true` only value.
 
 ## 0.9.1 / 2017-07-06